+++ Sudoku.c
@@ -4,7 +4,6 @@
 #include <string.h>
 
 /*----- Global Declarations ----- */
-int tempGrid[9][9];
 int grid[9][9];
 char flag,flag2;
 
@@ -17,7 +16,6 @@ int getCandidates(int, int, int * );
 void getCol(int, int *);
 void getBox(int , int , int *);
 int getNextBlank(int row,int col, int *nextPos);
-void setTempGrid();
 void intro();
 
 /* ----- Start of Main ----- */
@@ -25,7 +23,6 @@ void intro();
 int main(void){
     intro();
     inputSudoku();
-    setTempGrid();
     printf("\n\n");
     solveSudoku();
     getchar();
@@ -33,7 +30,6 @@ int main(void){
 }
 /* ----- Populate the Sudoku Grid with Zero's ----- */
 void populateGrid(void){
-
 	int i=0,j=0;
 	for (i=0;i<9;++i) { 
 		for (j=0;j<9;++j) {
@@ -46,13 +42,15 @@ void output(){
 	for (i=0; i< 9 ;++i){
 		printf("\n");
 		if (i == 0 || i == 3 || i == 6 ){
-			printf("  --------------------------------------------------  \n");
+			printf(" -------------------------------------------------  \n");
 		}
 		for (j=0;j < 9 ;++j){
-			if (j==2 || j==5 || j==8) printf("  %d  |", tempGrid[i][j]);		
-			else printf("  %d   ", tempGrid[i][j]);				 
+			if (j==2 || j==5 || j==8) printf("  %d  |", grid[i][j]);		
+			else if (j==0) printf(" | %d   ", grid[i][j]);
+			else printf(" %d   ", grid[i][j]);			 
 		}
 	}
+	printf("\n -------------------------------------------------  \n");
 }
 void inputSudoku(void){
 	int i = 0, j=0;
@@ -65,8 +63,7 @@ void inputSudoku(void){
 			for (j=0;j<9;++j){
 				grid[i][j] = str[j] - 48; /* converts ascii to integer */
 				str[j] = ' ';
-			} /* converts ascii to integer */
-			
+			} 			
 			if (feof (puzzle)){break;}
 		}
 		fclose(puzzle);
@@ -74,9 +71,8 @@ void inputSudoku(void){
 	else perror("Puzzle.txt");				
 }
 int getCandidates(int row, int col, int* cands){
-
 	int i = 0, j = 0, k = 0,array[9] = {0,0,0,0,0,0,0,0,0}, tempCol[9] = {0,0,0,0,0,0,0,0,0}, tempBox[9] = {0,0,0,0,0,0,0,0,0}, tempRow[9] = {0,0,0,0,0,0,0,0,0};
-	memcpy(tempRow,tempGrid[row],sizeof(tempRow));
+	memcpy(tempRow,grid[row],sizeof(tempRow));
 	getCol(col, tempCol);
 	getBox(row,col,tempBox);
 
@@ -99,7 +95,7 @@ int getCandidates(int row, int col, int* cands){
 }				
 void getCol(int numberOfCol, int *col){
 	int i;
-	for (i=0;i<9;++i){col[i] = tempGrid[i][numberOfCol];}
+	for (i=0;i<9;++i){col[i] = grid[i][numberOfCol];}
 	}
 void getBox(int row, int col, int *tempBlock){
 	int i,j;
@@ -109,15 +105,15 @@ void getBox(int row, int col, int *tempBlock){
 	if (col<3){j=1;}
 	else if (col <6 && col > 2){j=4;}
 	else {j = 7;}
-	tempBlock[0] = tempGrid[i-1][j-1];
-	tempBlock[1] = tempGrid[i][j-1];
-	tempBlock[2] = tempGrid[i+1][j-1];
-	tempBlock[3] = tempGrid[i-1][j];
-	tempBlock[4] = tempGrid[i][j];
-	tempBlock[5] = tempGrid[i+1][j];
-	tempBlock[6] = tempGrid[i-1][j+1];
-	tempBlock[7] = tempGrid[i][j+1];
-	tempBlock[8] = tempGrid[i+1][j+1];		
+	tempBlock[0] = grid[i-1][j-1];
+	tempBlock[1] = grid[i][j-1];
+	tempBlock[2] = grid[i+1][j-1];
+	tempBlock[3] = grid[i-1][j];
+	tempBlock[4] = grid[i][j];
+	tempBlock[5] = grid[i+1][j];
+	tempBlock[6] = grid[i-1][j+1];
+	tempBlock[7] = grid[i][j+1];
+	tempBlock[8] = grid[i+1][j+1];		
 }
 int bruteForce(int row,int col){
 	// ----- Declarations -----
@@ -125,21 +121,21 @@ int bruteForce(int row,int col){
 	int cands[9] = {0,0,0,0,0,0,0,0,0};
 	// ------------------------
 	candFlag = getCandidates(row,col,cands); //get possible candidates for row and col.
-	if (candFlag == 0){tempGrid[row][col] = 0;return 0;} // checking if candidates is null. if null, return 0 (backtracking)
+	if (candFlag == 0){grid[row][col] = 0;return 0;} // checking if candidates is null. if null, return 0 (backtracking)
 	blankFlag = getNextBlank(row,col,nextPos); //getting next square with a 0 (blank)
 	if (blankFlag == 0){ // if no next blanks, set current box to the only candidate (as it can only be one), then return 1 (solved)
-		tempGrid[row][col] = cands[0];
+		grid[row][col] = cands[0];
 		return 1;
 	}
 	for (i=0;i < 9;++i){
 		
-		tempGrid[row][col] = cands[i];
+		grid[row][col] = cands[i];
 		flag = bruteForce(nextPos[0],nextPos[1]);
 		if (flag == 1){return 1;}
-		if (cands[i + 1] == 0){tempGrid[row][col] = 0;return 0;}
+		if (cands[i + 1] == 0){grid[row][col] = 0;return 0;}
 		}
 		
-	tempGrid[row][col] = 0;
+	grid[row][col] = 0;
 	return 0;
 	}
 int getNextBlank(int row,int col, int *nextPos){
@@ -148,7 +144,7 @@ int getNextBlank(int row,int col, int *nextPos){
 	while (i<9){
 		while (j < 8){
 			j++;
-			if (tempGrid[i][j] == 0){
+			if (grid[i][j] == 0){
 				nextPos[0] = i;
 				nextPos[1] = j; 
 				return 1;
@@ -163,23 +159,20 @@ int getNextBlank(int row,int col, int *nextPos){
 	}
 void solveSudoku(){
 	int solvedFlag = 0,firstPos[2] = {0,0};
-	setTempGrid();
-	getNextBlank( 0 , -1 , firstPos);
-	
+	getNextBlank( 0 , -1 , firstPos);	
 	solvedFlag = bruteForce(firstPos[0],firstPos[1]);
 	
 	if (solvedFlag ==1){
+		printf("The solution: \n");
 		output();
 		printf("\n\nThe sudoku has been solved.");
 	}
 }
-void setTempGrid(){
-	memcpy(tempGrid,grid,sizeof(tempGrid));
-		}
+
 void intro(void){
-	printf("******************************************* \n");
+	printf("*********************************************** \n");
 	printf("    Welcome to Sudoku Solver Version 1.1.0\n");
-	printf("******************************************* \n\n\n");	
+	printf("*********************************************** \n");	
 	printf("Please have your puzzle file inside the same folder as the program.\n");
 	
 	}